Separate component and dumb component
