+++ src/components/common/Frame.tsx
@@ -3,8 +3,7 @@ import React, { CSSProperties } from 'react'
 const CSS_DEFAULT: CSSProperties = {
   margin: '2rem auto',
   border: '2px dashed gray',
-  padding: '2rem',
-  minHeight: '20rem',
+  padding: '1.5rem',
 }
 
 interface Props {
